Fix DW1000 IRQ configuration.
The previous commit e23eebde no longer uses the DFR status flag.
This commit updates the IRQ configuration to use the FCG flag in
place of the DFR flag.

This change fixes the DecaDriver not getting notified when the
receiver is disabled (due to a frame being received) and causing a
user application waiting on the DecaDriver.Receiver forever.