Perform a soft reset of the receiver when certain rx errors occur.
Section 4.1.6 of the DW1000 User Manual explains that there is an
issue with the re-initialization of the receiver following certain
types of rx errors, and that the receiver must be reset to ensure
correct timestamp operation for the next received packet.

This commit updates the DecaDriver IRQ handler to perform this
procedure.