Refactor and split up the DecaDriver.
This change separates the Driver, Receiver, and Transmitter
protected objects into their own child packages under the DecaDriver
package. The new child packages are:
  * DecaDriver.Core handles the main Driver PO.
  * DecaDriver.Tx handles the Transmitter driver PO.
  * DecaDriver.Rx handles the Receiver driver PO.

This change has been make primarily as a workaround for a bug in
GNATprove (related to the Attach_Handler aspect in the DW1000_IRQ
procedure) which causes GNATprove to crash. Separating the
DecaDriver into different packages allows GNATprove to be run
without crashing on the majority of the DecaDriver code in order to
minimize the amount of code that is not formally verified.

The secondary reason for splitting up the DecaDriver is to aid with
maintenance and readability, as responsibilities for transmitting,
receiving, and configuration are no longer all mixed up into the
same package.