Detect receiver errors by default.
This change updates the DecaDriver to detect all packet reception
errors by default (the previous default was that no errors are
detected).

This is a more sensible default since the DW1000 disables its
receiver by default when an error occurs, which combined with the
previous driver defaults means that it is possible that the user
may wait forever for a packet if they did not reconfigure the
driver to detect errors.