Move the DecaDriver's Tx_Power_Config_Type to DW1000.Driver
This change replaces the DW1000.Driver.Tx_Power_Config_Type with
the definition in DecaDriver.Tx.Tx_Power_Config_Type so that the
EVB1000_Tx_Power package is not dependent on the DecaDriver.

Not all applications are expected to use the DecaDriver (they might
want to make their own application-specific driver, based on
DW1000.Driver).