Reduce maximum Byte_Array length to avoid overflow with 'Length attribute
When a subprogram takes an unconstrained Byte_Array as a parameter and
references its 'Length attribute it is possible that in the case of the
largest possible array ('First = 0 and 'Last = 2**31 - 1) then 'Length
is equal to 2**31 which is 1 past the maximum range supported by the
Natural type (on a 32-bit ARM microcontroller, at least) which is now
detected by SPARK Discovery 2017. This commit constrains the index range
of the Byte_Array type by 1 to allow all possible 'Length values to fit
inside the Natural type.