Make DW1000.BSP a synchronous interface.
This adds the Synchronous attribute to the abstract
DW1000.BSP.Device_State, and realises this in the DWM1001
implementation with a protected object.

The purpose of this change is to handle flow analysis with GNATprove,
which detects that DW1000.BSP.Device_State is accessed by both
an interrupt (DecaDriver.IRQ_Handler) and the environment task, and
issues a warning.
This problem is resolved by making Dw1000.BSP.Device_State Synchronous.

We still need to support ZFP runtimes, so we can't use a protected object
directly in the package spec of DW1000.BSP. A ZFP implementation of
DW1000.BSP would instead protect against race conditions by disabling
the appropriate interrupt(s) which could use the SPI.