Refactor non-SPARK code in DW1000.System_Time to be in SPARK.
This commit refactors the conversion subprograms in the
DW1000.System_Time package so that they are written entirely in
SPARK.

The primary benefit of this refactoring is so that the definitions
of the conversion subprograms are available to GNATprove, which
can provide some useful properties for proofs e.g. that converting
a Bits_40 value to Fine_System_Time and back again yields the
original value.