Refactor distance types in the DW1000.Ranging package.
This refactors the `Distance` fixed-point type into a `Meters`
floating-point type.

A float is used instead of a fixed-point type as a trade-off
to avoid difficulties dealing with conversions between different
fixed-point types or dealing with fixed-point types with a
fractional 'Small.

To elaborate, if `Meters` was a fixed-point type then in order for
the conversion from `System_Time_Span` to a distance in `Meters`
to be supported, the 'Delta and 'Small of the `Meters` type would
need to be an integer multiple of `System_Time_Span'Delta`.
Otherwise, the result of the fixed-point conversion would not
always belong to the perfect result set as defined in Ada RM G.2.3.
This is not a problem in itself, but one of GNATprove's limitations
is that it only supports fixed-point operations that belong to the
perfect result set, otherwise the conversion is not provable in SPARK.

The problem that arises from solving the aforementioned fixed-point
to fixed-point conversion problem is that the resulting definition
of `Meters` wouild have a fractional 'Small. Such types cause other
difficulties, such as that integer values (e.g. 1.0, 3.0, etc...)
cannot be represented exactly.

So although having a fixed-point type for `Meters` would be more
efficient (particularly for MCUs without an FPU), having
a floating-point type is more convenient to use.