Refactor the DecaDriver into a single package. (#5)
The purpose of these changes is to simplify the DecaDriver to just the set of functionality needed to manage the transmission and reception of data frames. The main change is to refactor the DecaDriver packages into a single package, providing a single point of access to handle all transmit & receive operations.

Part of the reason for the original DecaDriver design was that the DW1000.BSP package was not synchronised; i.e. there was no protection against race conditions between tasks and/or interrupts to the SPI interface. The DecaDriver was intended to serve as the synchronisation point with a protected object. Now that DW1000.BSP is synchronised this is not needed, so the DecaDriver only needs to protect its internal state from race conditions.

This means that the various functions provided by DW1000.Driver can be called directly, rather than needing synchronisation via the DecaDriver. Of course, this only protects the SPI interface from race conditions - it does not protect against cases where the DW1000 is reconfigured during an ongoing transmit/receive operation as these are asynchronous.